+++ src/main/java/com/onek/dao/CritereDao.java
@@ -1,0 +1,10 @@
+package com.onek.dao;
+
+import java.util.List;
+
+import com.onek.model.Critere;
+
+public interface CritereDao {
+	void addCriteres(List<Critere> criteres);
+	Critere findById(Integer id);
+}
