+++ src/main/java/com/onek/dao/EvenementDaoImpl.java
@@ -1,6 +1,7 @@
 package com.onek.dao;
 
 import java.io.Serializable;
+import java.util.ArrayList;
 import java.util.List;
 
 import org.hibernate.Hibernate;
@@ -15,64 +16,62 @@ import com.onek.model.Evenement;
 @Repository
 public class EvenementDaoImpl implements EvenementDao, Serializable {
 	private static final long serialVersionUID = 1L;
-	
+
 	@Autowired
 	private SessionFactory sessionFactory;
 
 	@Override
 	public void addEvenement(Evenement event) {
-		
 		Session session = sessionFactory.openSession();
 		session.beginTransaction();
 		session.save(event);
 		session.getTransaction().commit();
 		session.close();
-
-		System.out.println("Add done");		
+		System.out.println("Add done !");
 	}
 
 	@Override
 	public Evenement findById(int id) {
-		Session session = sessionFactory.openSession();		
-        session.beginTransaction();
-        
+		Session session = sessionFactory.openSession();
+		session.beginTransaction();
 		Evenement event = (Evenement) session.createQuery("FROM Evenement WHERE idevent = :id").setParameter("id", id)
 				.getSingleResult();
-        Hibernate.initialize(event.getCriteres());        
-        Hibernate.initialize(event.getCandidats());
-        List<Critere> criteres = event.getCriteres();
-        for(Critere c : criteres) {
-        	Hibernate.initialize(c.getDescripteurs());
-        }    
-        session.getTransaction().commit();
-        session.close();
-        
-        System.out.println("Find done - Id: " + event.getIdevent());
-		return event;		
-	}	
-	
+		Hibernate.initialize(event.getCriteres());
+		Hibernate.initialize(event.getCandidats());
+		List<Critere> criteres = event.getCriteres();
+		for (Critere c : criteres) {
+			Hibernate.initialize(c.getDescripteurs());
+		}
+		session.getTransaction().commit();
+		session.close();
+		System.out.println("Find done - Id: " + event.getIdevent());
+		return event;
+	}
+
+	@SuppressWarnings("unchecked")
 	@Override
 	public List<Evenement> findByIdUser(int idUser) {
-		Session session = sessionFactory.openSession();		
-        session.beginTransaction();
-        @SuppressWarnings("unchecked")
-		List<Evenement> events = (List<Evenement>) session.createQuery("FROM Evenement WHERE iduser = :iduser").setParameter("iduser", idUser).list();
-        session.getTransaction().commit();
-        session.close();
-        return events;		
+		Session session = sessionFactory.openSession();
+		session.beginTransaction();
+		List<Evenement> events = (List<Evenement>) session
+				.createQuery("FROM Evenement WHERE (isdeleted = :isDeleted AND iduser = :iduser)")
+				.setParameter("iduser", idUser).setParameter("isDeleted", false).list();
+		session.getTransaction().commit();
+		session.close();
+		return events;
 	}
-	
-	@Override
 
+	@Override
 	public Evenement findByCode(String code) {
-		Session session = sessionFactory.openSession();		
-        session.beginTransaction();
-        Evenement event = (Evenement) session.createQuery("FROM Evenement WHERE code = :code").setParameter("code", code).getSingleResult();
-        session.getTransaction().commit();
-        session.close();
-        return event;
+		Session session = sessionFactory.openSession();
+		session.beginTransaction();
+		Evenement event = (Evenement) session.createQuery("FROM Evenement WHERE code = :code")
+				.setParameter("code", code).getSingleResult();
+		session.getTransaction().commit();
+		session.close();
+		return event;
 	}
-	
+
 	@Override
 	public void editEvenement(Evenement event) {
 		Session session = sessionFactory.openSession();
@@ -80,5 +79,28 @@ public class EvenementDaoImpl implements EvenementDao, Serializable {
 		session.update(event);
 		session.getTransaction().commit();
 		session.close();
+	}
+	
+	@SuppressWarnings("unchecked")
+	@Override
+	public List<Evenement> findAll() {
+		List<Evenement> events = new ArrayList<>();		
+		Session session = sessionFactory.openSession();
+		session.beginTransaction();	
+		events = (List<Evenement>) session.createQuery("from Evenement").list();
+		session.getTransaction().commit();
+		session.close();
+		return events;
+	}
+	
+	public void supprimerEvent(int idevent) {
+		Session session = sessionFactory.openSession();
+		session.beginTransaction();
+		Evenement event = (Evenement) session.createQuery("from Evenement where idevent = :idevent").setParameter("idevent", idevent).getSingleResult();
+		event.setIsdeleted(true);
+		session.update(event);
+		session.getTransaction().commit();
+		session.close();
+		System.out.println("Delete done (isDeleted = true) - Id: " + event.getIdevent());
 	}
 }
