--- src/main/java/com/onek/dao/GrilleDao.java
@@ -1,10 +1,0 @@
-package com.onek.dao;
-
-import java.util.List;
-
-import com.onek.model.Critere;
-
-public interface GrilleDao {
-	void addCriteres(List<Critere> criteres);
-	Critere findById(Integer id);
-}
