--- src/main/java/com/onek/service/AccueilService.java
@@ -1,11 +1,0 @@
-package com.onek.service;
-
-import java.util.List;
-
-import com.onek.model.Evenement;
-
-public interface AccueilService {
-	public List<Evenement> listEvents();
-	public List<Evenement> myListEvents(int iduser);
-	public void supprimerEvent(int idevent);
-}
