--- src/main/java/com/onek/service/AccueilServiceImpl.java
@@ -1,30 +1,0 @@
-package com.onek.service;
-
-import java.io.Serializable;
-import java.util.List;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import com.onek.dao.AccueilDao;
-import com.onek.model.Evenement;
-@Service
-public class AccueilServiceImpl implements AccueilService, Serializable{
-	private static final long serialVersionUID = 1L;
-	@Autowired
-	private AccueilDao accueildao;
-	
-	@Override
-	public List<Evenement> listEvents(){
-		return accueildao.listEvents();
-	}
-	@Override
-	public void supprimerEvent(int idevent) {
-		accueildao.supprimerEvent(idevent);
-	}
-	@Override
-	public List<Evenement> myListEvents(int iduser) {
-		
-		return accueildao.myListEvents(iduser);
-	}
-}
