+++ src/main/java/com/onek/service/EvenementService.java
@@ -1,5 +1,7 @@
 package com.onek.service;
 
+import java.util.List;
+
 import com.onek.model.Evenement;
 
 public interface EvenementService {
@@ -7,4 +9,7 @@ public interface EvenementService {
 	public boolean isValid(Evenement event);
 	public Evenement findById(int id);
 	public void editEvenement(Evenement event);
+	public List<Evenement> findAll();
+	public void supprimerEvent(int idevent);
+	public List<Evenement> myListEvents(int iduser);
 }
