+++ src/main/java/com/onek/service/EvenementServiceImpl.java
@@ -1,6 +1,7 @@
 package com.onek.service;
 
 import java.io.Serializable;
+import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -13,7 +14,7 @@ public class EvenementServiceImpl implements EvenementService, Serializable{
 	private static final long serialVersionUID = 1L;
 	
 	@Autowired
-	EvenementDao eventDao;
+	private EvenementDao eventDao;
 	
 	@Override
 	public void addEvenement(Evenement event) {
@@ -33,5 +34,20 @@ public class EvenementServiceImpl implements EvenementService, Serializable{
 	@Override
 	public void editEvenement(Evenement event) {
 		eventDao.editEvenement(event);
+	}
+	
+	@Override
+	public List<Evenement> findAll(){
+		return eventDao.findAll();
+	}
+	
+	@Override
+	public void supprimerEvent(int idevent) {
+		eventDao.supprimerEvent(idevent);
+	}
+	
+	@Override
+	public List<Evenement> myListEvents(int iduser) {		
+		return eventDao.findByIdUser(iduser);
 	}
 }
