+++ src/main/java/com/onek/service/GrilleServiceImpl.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.onek.dao.GrilleDao;
+import com.onek.dao.CritereDao;
 import com.onek.model.Critere;
 
 @Service
@@ -14,10 +14,10 @@ public class GrilleServiceImpl implements GrilleService, Serializable{
 	private static final long serialVersionUID = 1L;
 	
 	@Autowired
-	GrilleDao grilleDao;
+	private CritereDao critereDao;
 	
 	@Override
 	public void addCriteres(List<Critere> criteres) {
-		grilleDao.addCriteres(criteres);		
+		critereDao.addCriteres(criteres);		
 	}
 }
